+++ src/drivers/display.h
@@ -26,8 +26,8 @@ public:
     virtual void refresh() = 0;
     virtual int clear();
     //widget functions
-    virtual void drawText(const std::string& text, const uint8_t& x, const uint8_t& y, bool isHighlighted = false) = 0;
-    virtual void drawBox(const uint8_t& width, const uint8_t& height, const bool hasBorder, const bool isFilled, const uint8_t& x, const uint8_t& y) = 0; 
+    virtual void drawText(const std::string& text, const uint16_t& x, const uint16_t& y, bool isHighlighted = false) = 0;
+    virtual void drawBox(const uint16_t& width, const uint16_t& height, const bool hasBorder, const bool isFilled, const uint16_t& x, const uint16_t& y) = 0; 
 };
 
 #endif // DISPLAY_H_
