+++ src/drivers/tty_display.cpp
@@ -22,10 +22,10 @@ int TtyDisplay::clear() {
     return 0;
 }
 
-void TtyDisplay::drawText(const std::string& text, const uint8_t& x, const uint8_t& y, bool isHighlighted) {
+void TtyDisplay::drawText(const std::string& text, const uint16_t& x, const uint16_t& y, bool isHighlighted) {
    std::cout << text << std::endl; 
 }
 
-void TtyDisplay::drawBox(const uint8_t& width, const uint8_t& height, const bool hasBorder, const bool isFilled, const uint8_t& x, const uint8_t& y) {
+void TtyDisplay::drawBox(const uint16_t& width, const uint16_t& height, const bool hasBorder, const bool isFilled, const uint16_t& x, const uint16_t& y) {
     spdlog::debug("draw Box at {} {}", x, y);
 }
