+++ src/drivers/waveshare_eink.cpp
@@ -70,7 +70,7 @@ int WaveshareEink::clear() {
     return 0;
 }
 
-void WaveshareEink::drawText(const std::string& text, const uint8_t& x, const uint8_t& y, bool isHighlighted) {
+void WaveshareEink::drawText(const std::string& text, const uint16_t& x, const uint16_t& y, bool isHighlighted) {
     spdlog::debug("draw Text at {} {}", x, y);
     std::lock_guard<std::mutex> lock(m_bufferMutex);
     Paint_SelectImage(m_frameBuffer);
@@ -82,7 +82,7 @@ void WaveshareEink::drawText(const std::string& text, const uint8_t& x, const ui
     //m_cursor.x += Font24.Width * text.length();
 }
 
-void WaveshareEink::drawBox(const uint8_t& width, const uint8_t& height, const bool hasBorder, const bool isFilled, const uint8_t& x, const uint8_t& y) {
+void WaveshareEink::drawBox(const uint16_t& width, const uint16_t& height, const bool hasBorder, const bool isFilled, const uint16_t& x, const uint16_t& y) {
     spdlog::debug("draw Box at {} {}", x, y);
     std::lock_guard<std::mutex> lock(m_bufferMutex);
     Paint_SelectImage(m_frameBuffer);
