+++ src/ui/time_widget.h
@@ -19,7 +19,7 @@ private:
 public:
 	TimeWidget(bool isHighlighted = false);
 	~TimeWidget();
-	virtual void render(Display* display, const uint8_t& x, const uint8_t& y) const override;
+	virtual void render(Display* display, const uint16_t& x, const uint16_t& y) const override;
 	virtual std::unique_ptr<Widget> clone() override;
 	void updateTime();
 	void updateThreadLoop();
